requirements.txt
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
pyasn1==0.6.1
PySimpleGUI==5.0.10
rsa==4.9.1
